pyproject.toml
[project]
name = "t2asciiart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "jinja2>=3.1.6",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
